need updates

+++ src/control_unit/control_unit_A/lora_communication.h
@@ -17,6 +17,7 @@
 
 #define localAddress 0xAA
 
+// Sets up the lora pins and frequencies
 void initializeLoRa();
 
 // Reads data from inertial unit and stores them in AzimuthElevation structure
@@ -26,11 +27,14 @@ AzimuthElevation* readFromInertialUnit();
 // Restarts inertial unit
 // @return 0 if success, -1 if error
 // @param azimuth value indicates calibration of azimuth on inertial unit
-  // if its -1 it stays the same and sensor just resets
-int restartInertialUnit(double azimuth = -1, double calibrationMatrix[3][3] = nullptr);
+// if its -1 it stays the same and sensor just resets
+int restartInertialUnit(double azimuth = -1, int calibrationMatrix[3][3] = nullptr);
 
+// Display given Azimuth and Elevation on the display
+// @param ae AzimuthElevation struct that should be displayed
 void displayAE(AzimuthElevation* ae);
 
+// Reads from senzor, evaluates danger, logs, displays on screen and sends to connected clients
 void doOperations();
 
 #endif // LORA_COMMUNICATION_H
